requirements.txt
# Core dependencies
pyyaml>=6.0
python-dateutil>=2.8.2

# Optional: semantic search
# sentence-transformers>=2.2.0
# torch>=2.0.0

# Optional: LLM integration
# openai>=1.0.0
